backend/requirements.txt
aerich
fastapi
uvicorn
aiomysql
tortoise-orm[asyncmy]
cryptography